refactor: merge selected loader module into core selected loader
  * move SelectedMapLoaderModule declaration/implementation into selected_map_loader hpp/cpp
  * remove obsolete selected_map_loader_module hpp/cpp files
  * update node and test includes to selected_map_loader.hpp
