+++ map/autoware_map_loader/src/pointcloud_map_loader/pointcloud_map_loader_node.hpp
@@ -18,7 +18,7 @@
 #include "differential_map_loader.hpp"
 #include "partial_map_loader.hpp"
 #include "pointcloud_map_loader.hpp"
-#include "selected_map_loader_module.hpp"
+#include "selected_map_loader.hpp"
 
 #include <rclcpp/rclcpp.hpp>
